+++ packages/web-vue/lib/option-list/index.vue
@@ -10,6 +10,7 @@ import { GROUP_OPTION_TYPE } from '../share/const'
 import { getCurrentInstance, useSlots, withScopeId } from 'vue'
 import VirtualList from '../virtual-list/index.vue'
 import type { JSX } from 'vue/jsx-runtime'
+import Scroll from '../scroll-bar/index.vue'
 
 defineOptions({
 	name: 'OptionList'
@@ -135,15 +136,31 @@ const renderList = () => {
 		})
 	return (
 		<ul class="px-option-list">
-			{!props.virtualScroll ? (
-				list.map((item) => item.el)
-			) : (
-				<VirtualList
-					class={'px-option-list-virtual-list'}
-					list={list.map((item) => ({ render: () => item.el, key: item.key }))}
-					{...props.virtualListProps}
-				></VirtualList>
-			)}
+			<Scroll class="px-option-list-scroll">
+				{!props.virtualScroll ? (
+					list.map((item) => item.el)
+				) : (
+					<VirtualList
+						class={'px-option-list-virtual-list'}
+						list={list.map((item) => ({ render: () => item.el, key: item.key }))}
+						{...props.virtualListProps}
+					>
+						{{
+							'scroll-container': ({
+								children,
+								onScroll
+							}: {
+								children: JSX.Element
+								onScroll: (event: Event) => void
+							}) => (
+								<Scroll class={'px-option-list-scroll'} onScroll={onScroll}>
+									{children}
+								</Scroll>
+							)
+						}}
+					</VirtualList>
+				)}
+			</Scroll>
 		</ul>
 	)
 }